+++ packages/instro-unstable/CHANGELOG.md
@@ -1,5 +1,31 @@
 # Changelog
 
+## [0.3.0](https://github.com/nominal-io/instro/compare/instro-unstable-v0.2.1...instro-unstable-v0.3.0) (2026-06-25)
+
+
+### ⚠ BREAKING CHANGES
+
+* 149 bug documentation for get channel does not match behavior ([#151](https://github.com/nominal-io/instro/issues/151))
+
+### Bug Fixes
+
+* 149 bug documentation for get channel does not match behavior ([#151](https://github.com/nominal-io/instro/issues/151)) ([f4f3966](https://github.com/nominal-io/instro/commit/f4f396686fab9b437143730954d0638d28dbbb6d))
+* closes [#149](https://github.com/nominal-io/instro/issues/149). First, updates documentation about timeout behavior. ([f4f3966](https://github.com/nominal-io/instro/commit/f4f396686fab9b437143730954d0638d28dbbb6d))
+
+## [0.2.1](https://github.com/nominal-io/instro/compare/instro-unstable-v0.2.0...instro-unstable-v0.2.1) (2026-06-24)
+
+
+### Bug Fixes
+
+* **ethernetip:** remove strings from Python EIP surface ([#128](https://github.com/nominal-io/instro/issues/128)) ([c133d81](https://github.com/nominal-io/instro/commit/c133d814a66b68775fb5a05114cf1ccac22aa466))
+
+## [0.2.0](https://github.com/nominal-io/instro/compare/instro-unstable-v0.1.0...instro-unstable-v0.2.0) (2026-06-22)
+
+
+### Features
+
+* **ethernetip:** reconnect after transient failures ([#100](https://github.com/nominal-io/instro/issues/100)) ([e9a35da](https://github.com/nominal-io/instro/commit/e9a35da96e619719cad008ee591d6be250343cbd))
+
 ## [0.1.0](https://github.com/nominal-io/instro/compare/instro-unstable-v0.0.1...instro-unstable-v0.1.0) (2026-06-17)